handle errors, async event, and abort
This patch introduces proper error handling instead of simply asserting. In most
cases we attempt to propagate the error back to the caller(s), where we either
post a completion with the appropriate status code and status code type set,
or failing entirely the controller in case of an internal error. Also, partially
handle async event and abort commands.

In cases where we're supposed to complete an async event request in order to
communicate an error we must fail the controller as we don't fully handle
async event request.

Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>
Change-Id: Ib111de6d43c5853b875b7c165f0ca8e99373dbab