[RFC] nvmf: also disconnect IO queue pairs when shutting down controller
Previously we only do that when CC.EN changes from 1 to 0, we can also
do that when shutting down the controller.  The function will do nothing
if the IO queue pairs are already disconnected.  This is required for
MUSER transport because the Guest kernel will consider the controller
as the PCI device not RDMA/TCP fabrics.

Change-Id: Iadbfe5bff72382cf9091db57bf9718fa1a1f4a6c
Signed-off-by: Changpeng Liu <changpeng.liu@intel.com>