[RFC] nvmf/muser: refactor the NVMe command call process
This patch mainly do same function renaming work, only one place
is different with original logic, we should not change the command
dptr pointer which filled by Guest, the orginal logic is for hacking
the Identify Controller data structure, for now we will change
the SGL support to false by using "req->data" directly.

The Guest may set both PRP1 and PRP2 fields, but only PRP1 is used, so
we also remove the assertion to PRP2.

Change-Id: I21fd8298691feb127e41282ae3901b988e346574
Signed-off-by: Changpeng Liu <changpeng.liu@intel.com>