nvmf/muser: Unify admin and I/O queue creation paths
Both types of queues can now wait to send a fabric connect
command until they are assigned to a poll group. This is
possible because the Unix domain socket transport has a
built-in mechanism to delay receiving requests (simply
don't call recv on the socket until ready).

This also eliminates the temporary, internal poll group
that was used previously which results in a very nice
simplification of the whole code base.

Signed-off-by: Ben Walker <benjamin.walker@intel.com>
Change-Id: I85cb01824d6f45f50c345fbb8ce6cd6cc488d3e5